Use new unsafe idom to replace use of deprecated reflect.SliceHeader
The relevant code is moved to build-tagged files depending
on Go version, eventually the older one can go away.
