+++ certtostore_windows_go121.go
@@ -1,0 +1,31 @@
+//go:build windows && go1.21
+// +build windows,go1.21
+
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package certtostore
+
+import (
+	"crypto/x509"
+
+	"unsafe"
+
+	"golang.org/x/sys/windows"
+)
+
+// certContextToX509 creates an x509.Certificate from a Windows cert context.
+func certContextToX509(ctx *windows.CertContext) (*x509.Certificate, error) {
+	return x509.ParseCertificate(append([]byte{}, unsafe.Slice(ctx.EncodedCert, ctx.Length)...))
+}
